requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
strawberry-graphql[fastapi]==0.284.1
asyncpg==0.30.0
python-dotenv==1.0.1
pydantic==2.10.3
